+++ src/main/java/repast/model/agrosuccess/AgroSuccessContextBuilder.java
@@ -18,6 +18,7 @@ import me.ajlane.geo.repast.colonisation.LandCoverColoniser;
 import me.ajlane.geo.repast.colonisation.csr.CompletelySpatiallyRandomColoniser;
 import me.ajlane.geo.repast.colonisation.csr.CompletelySpatiallyRandomParams;
 import me.ajlane.geo.repast.fire.DefaultFireManager;
+import me.ajlane.geo.repast.fire.DefaultFireSpreader;
 import me.ajlane.geo.repast.fire.DefaultFlammabilityChecker;
 import me.ajlane.geo.repast.fire.FireManager;
 import me.ajlane.geo.repast.fire.FireParams;
@@ -331,7 +332,7 @@ public class AgroSuccessContextBuilder implements ContextBuilder<Object> {
         * (climateData.getMeanAnnualTemperature() / climateData.getTotalAnnualPrecipitation());
     double vegetationMoistureParam = meanNumFires; // lambda parameterises fuel moisture as well as
                                                    // number of fires
-    FireSpreader fireSpreader = new FireSpreader(lctLayer, fireCount, srCalc, wrCalc,
+    FireSpreader fireSpreader = new DefaultFireSpreader(lctLayer, fireCount, srCalc, wrCalc,
         lcfGetter.getMap(), windData.getWindDirectionProb(), windData.getWindSpeedProb());
     FlammabilityChecker<GridPoint> flamChecker = new DefaultFlammabilityChecker(lctLayer);
     return new DefaultFireManager(fireSpreader, flamChecker, lctLayer.getDimensions(),